+++ app/models/import_log.rb
@@ -1,0 +1,2 @@
+class ImportLog < ApplicationRecord
+end