+++ lib/tasks/migration/services/import_service.rb
@@ -24,25 +24,25 @@ module Migration
         import_record = nil
 
         begin
-          item = DigitoolItem.new({ :pid => pid}) if pid
+          item = DigitoolItem.new({ :pid => @pid}) if @pid.present?
           ##Get the dctypes
           dc_types = item.metadata_hash["type"] if !item.is_waiver?
           dc_types = dc_types.map(&:inspect).join(', ') if dc_types.kind_of?(Array)
-      
+
           # Get the lc_code
           lc_code = item.metadata_hash["localcollectioncode"]
-      
+
           #Determine the worktype from dc:type and lc_code
           work_type = MigrationHelper::get_worktype(dc_types, lc_code)
           # Determine the samvera collection from the worktype and lc_code
           collection_id = MigrationHelper::get_samvera_collection_id(work_type, lc_code) if work_type.present?
-          
+
           # get the migration config
           migration_config = MigrationHelper::get_migration_config(collection_id) if collection_id.present?
-          
+
           # lets create the tmp file location if it does not exist
           FileUtils::mkdir_p migration_config['tmp_file_location'] if migration_config.present?
-          
+
 
           # empty the item
           item = nil
@@ -57,9 +57,10 @@ module Migration
           # Return the following info, work_id, collection_id, title, pid, work_type
           import_record = {work_id: work.id,
                            collection_id: collection_id,
-                           title: work.title.first, worktype: work_type }
+                           digitool_collection_code: lc_code,
+                           pid: @pid,
+                           title: work.title.first, work_type: work_type }
           # Maybe we can add the import_record to db
-
 
         rescue StandardError => e
           raise e if count > 1
@@ -67,7 +68,7 @@ module Migration
           @logger.info "Failed importing #{@pid} #{count} times, retrying. Error: #{e.message}"
           return import(count: count)
         end
- 
+
         import_record
 
       end